requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
pillow==10.0.0
requests==2.31.0
